Fix the --url option
This option appears to be intended to allow the Rack app to be mounted
at a specific path, rather than /, but it doesn't actually do anything.
The :Path option is set and passed along to the Rack::Handler, but none
of the built-in handlers appear to support it.

So this patch fixes it by wrapping the Shotgun::Loader app in a `mount`
block, which defaults to "/", but can be set with the --url/-u option.
I didn't add any tests, since all of this stuff happens in bin/shotgun.